Refactor for new theme division
style_* + color_* = theme_*
